Refactor to use n8n as ingest trigger
